Add --all functions for managing child hotkeys and take.
Implemented functions to set, revoke, and manage children hotkeys within the staking module. This includes functionalities for proportion assignment, querying childkey take, and displaying current childkey configurations across different subnets.
