+++ bittensor_cli/cli.py
@@ -16,8 +16,9 @@ from rich.table import Column, Table
 from .src import HYPERPARAMS, defaults, HELP_PANELS
 from bittensor_cli.src.bittensor import utils
 from .src.bittensor.async_substrate_interface import SubstrateRequestException
-from .src.commands import root, stake, subnets, sudo, wallets
+from .src.commands import root, subnets, sudo, wallets
 from .src.commands import weights as weights_cmds
+from .src.commands.stake import children_hotkeys, stake
 from bittensor_cli.src.bittensor.subtensor_interface import SubtensorInterface
 from bittensor_cli.src.bittensor.utils import (
     console,
@@ -3079,7 +3080,7 @@ class CLIManager:
                 "Enter netuid (leave blank for all)", default=None, show_default=True
             )
         return self._run_command(
-            stake.get_children(wallet, self.initialize_chain(network, chain), netuid)
+            children_hotkeys.get_children(wallet, self.initialize_chain(network, chain), netuid)
         )
 
     def stake_set_children(
@@ -3092,7 +3093,17 @@ class CLIManager:
         wallet_path: str = Options.wallet_path,
         network: str = Options.network,
         chain: str = Options.chain,
-        netuid: int = Options.netuid,
+        netuid: Optional[int] = typer.Option(
+            None,
+            help="The netuid (network unique identifier) of the subnet within the root network, (e.g. 1)",
+            prompt=False,
+        ),
+        all_netuids: bool = typer.Option(
+            False,
+            "--all-netuids",
+            "--all",
+            help="When this flag is used it sets children on all subnets on the bittensor network.",
+        ),
         proportions: list[float] = typer.Option(
             [],
             "--proportions",
@@ -3127,6 +3138,15 @@ class CLIManager:
         neurons (hotkeys) on the network. It allows for a strategic allocation of authority to enhance network participation and influence.
         """
         self.verbosity_handler(quiet, verbose)
+        if all_netuids and netuid:
+            err_console.print("Specify either netuid or all, not both.")
+            raise typer.Exit()
+        if all_netuids:
+            netuid = None
+        elif not netuid:
+            netuid = IntPrompt.ask(
+                "Enter netuid (leave blank for all)", default=None, show_default=True
+            )
         children = list_prompt(children, str, "Enter the child hotkeys (ss58)")
         proportions = list_prompt(
             proportions,
@@ -3141,14 +3161,14 @@ class CLIManager:
             raise typer.Exit()
         wallet = self.wallet_ask(wallet_name, wallet_path, wallet_hotkey)
         return self._run_command(
-            stake.set_children(
-                wallet,
-                self.initialize_chain(network, chain),
-                netuid,
-                children,
-                proportions,
-                wait_for_finalization,
-                wait_for_inclusion,
+            children_hotkeys.set_children(
+                wallet=wallet,
+                subtensor=self.initialize_chain(network, chain),
+                netuid=netuid,
+                children=children,
+                proportions=proportions,
+                wait_for_finalization=wait_for_finalization,
+                wait_for_inclusion=wait_for_inclusion,
             )
         )
 
@@ -3159,7 +3179,17 @@ class CLIManager:
         wallet_path: Optional[str] = Options.wallet_path,
         network: Optional[str] = Options.network,
         chain: Optional[str] = Options.chain,
-        netuid: int = Options.netuid,
+        netuid: Optional[int] = typer.Option(
+            None,
+            help="The netuid (network unique identifier) of the subnet within the root network, (e.g. 1)",
+            prompt=False,
+        ),
+        all_netuids: bool = typer.Option(
+            False,
+            "--all-netuids",
+            "--all",
+            help="When this flag is used it sets children on all subnets on the bittensor network.",
+        ),
         wait_for_inclusion: bool = Options.wait_for_inclusion,
         wait_for_finalization: bool = Options.wait_for_finalization,
         quiet: bool = Options.quiet,
@@ -3187,8 +3217,17 @@ class CLIManager:
         """
         self.verbosity_handler(quiet, verbose)
         wallet = self.wallet_ask(wallet_name, wallet_path, wallet_hotkey)
+        if all_netuids and netuid:
+            err_console.print("Specify either netuid or all, not both.")
+            raise typer.Exit()
+        if all_netuids:
+            netuid = None
+        elif not netuid:
+            netuid = IntPrompt.ask(
+                "Enter netuid (leave blank for all)", default=None, show_default=True
+            )
         return self._run_command(
-            stake.revoke_children(
+            children_hotkeys.revoke_children(
                 wallet,
                 self.initialize_chain(network, chain),
                 netuid,
@@ -3204,10 +3243,18 @@ class CLIManager:
         wallet_path: Optional[str] = Options.wallet_path,
         network: Optional[str] = Options.network,
         chain: Optional[str] = Options.chain,
-        netuid: int = Options.netuid,
-        wait_for_inclusion: bool = Options.wait_for_inclusion,
-        wait_for_finalization: bool = Options.wait_for_finalization,
-        prompt: bool = Options.prompt,
+        hotkey: Optional[str] = None,
+        netuid: Optional[int] = typer.Option(
+            None,
+            help="The netuid (network unique identifier) of the subnet within the root network, (e.g. 1)",
+            prompt=False,
+        ),
+        all_netuids: bool = typer.Option(
+            False,
+            "--all-netuids",
+            "--all",
+            help="When this flag is used it sets children on all subnets on the bittensor network.",
+        ),
         take: Optional[float] = typer.Option(
             None,
             "--take",
@@ -3215,6 +3262,9 @@ class CLIManager:
             help="Enter take for your child hotkey",
             prompt=False,
         ),
+        wait_for_inclusion: bool = Options.wait_for_inclusion,
+        wait_for_finalization: bool = Options.wait_for_finalization,
+        prompt: bool = Options.prompt,
         quiet: bool = Options.quiet,
         verbose: bool = Options.verbose,
     ):
@@ -3240,12 +3290,22 @@ class CLIManager:
         """
         self.verbosity_handler(quiet, verbose)
         wallet = self.wallet_ask(wallet_name, wallet_path, wallet_hotkey)
+        if all_netuids and netuid:
+            err_console.print("Specify either netuid or all, not both.")
+            raise typer.Exit()
+        if all_netuids:
+            netuid = None
+        elif not netuid:
+            netuid = IntPrompt.ask(
+                "Enter netuid (leave blank for all)", default=None, show_default=True
+            )
         return self._run_command(
-            stake.childkey_take(
+            children_hotkeys.childkey_take(
                 wallet=wallet,
                 subtensor=self.initialize_chain(network, chain),
                 netuid=netuid,
                 take=take,
+                hotkey=hotkey,
                 wait_for_inclusion=wait_for_inclusion,
                 wait_for_finalization=wait_for_finalization,
                 prompt=prompt,
